+++ util-app/src/main/scala/com/twitter/app/ClassPath.scala
@@ -60,7 +60,7 @@ private[app] sealed abstract class ClassPath[CpInfo <: ClassPath.Info] {
       browseUri0(uri, loader, buf, seenUris)
       seenUris += uri
     }
-    buf.result
+    buf.result()
   }
 
   // Note: In JDK 9+ URLClassLoader is no longer the default ClassLoader.
@@ -71,7 +71,7 @@ private[app] sealed abstract class ClassPath[CpInfo <: ClassPath.Info] {
   // TODO - add suppport for the ModulePath after dropping JDK 8 support.
   private[this] def urlsFromClasspath(): Array[URL] = {
     val classpath: String = System.getProperty("java.class.path")
-    classpath.split(File.pathSeparator).map { pathEntry: String =>
+    classpath.split(File.pathSeparator).map { (pathEntry: String) =>
       Paths.get(pathEntry).toAbsolutePath().toUri().toURL
     }
   }