+++ util-app/src/main/scala/com/twitter/app/Flags.scala
@@ -4,6 +4,7 @@ import scala.collection.immutable.TreeSet
 import scala.collection.mutable
 import scala.collection.mutable.ArrayBuffer
 import scala.jdk.CollectionConverters._
+import scala.reflect.ClassTag
 import scala.util.control.NonFatal
 
 /**
@@ -277,7 +278,7 @@ class Flags(argv0: String, includeGlobal: Boolean, failFastUntilParsed: Boolean)
    * @param name The name of the flag.
    * @param help The help string of the flag.
    */
-  def apply[T](name: String, help: String)(implicit _f: Flaggable[T], m: Manifest[T]): Flag[T] = {
+  def apply[T](name: String, help: String)(implicit _f: Flaggable[T], m: ClassTag[T]): Flag[T] = {
     val f = new Flag[T](name, help, m.toString, failFastUntilParsed)
     add(f)
     f